+++ chipnomad_lib/project.h
@@ -195,6 +195,7 @@ typedef struct Project {
   PitchTable pitchTable;
 
   uint16_t song[PROJECT_MAX_LENGTH][PROJECT_MAX_TRACKS];
+  uint8_t songHighlight[PROJECT_MAX_LENGTH][PROJECT_MAX_TRACKS];
   Chain chains[PROJECT_MAX_CHAINS];
   Phrase phrases[PROJECT_MAX_PHRASES];
   Groove grooves[PROJECT_MAX_GROOVES];